+++ src/Jute.jl
@@ -25,6 +25,9 @@ export @test_broken
 export @test_skip
 export @test_result
 export @test_fail
+export @inferred
+export @test_warn
+export @test_nowarn
 
 include("reporting.jl")